backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
channels>=4.0
daphne>=4.0
python-dotenv>=1.0
